Revert "move nightly and detector builds from oryx linux pool to ado pool (#1064)"
This reverts commit d5b2ebd60c18340e05e4483826afc957089b8b70.
