+++ src/BuildScriptGenerator/DotNetCore/DotNetCoreSdkVersions.cs
@@ -9,7 +9,7 @@ namespace Microsoft.Oryx.BuildScriptGenerator.DotNetCore
         public const string DotNetCore22SdkVersion = "2.2.207";
         public const string DotNetCore30SdkVersion = "3.0.103";
         public const string DotNetCore31SdkVersion = "3.1.412";
-        public const string DotNetCore50SdkVersion = "5.0.400";
+        public const string DotNetCore50SdkVersion = "5.0.302";
         public const string DotNet50SdkVersion = "5.0.400";
         public const string DotNet60SdkVersion = "6.0.100-rc.1.21458.32";
     }
