+++ vsts/pipelines/validation-detector.yml
@@ -9,12 +9,5 @@ pr:
     - tests/Detector.Tests/*
     - vsts/nugetpackages.signproj
     - vsts/pipelines/validation-detector.yml
-
-jobs:
-- job: Job_Detector
-  displayName: Detector Validation
-  pool:
-    vmImage: 'ubuntu-latest'
-  timeoutInMinutes: 480
-  steps:
-  - template: templates/_buildTemplateDetector.yml
+steps:
+- template: templates/_buildTemplateDetector.yml
